+++ CHANGELOG.md
@@ -25,7 +25,6 @@ If you have been using Makoto and/or Roboto directly, you were never at risk
         * In-memory, static, non-caching
     * Provide fully-fledged Android revocation lists
         * The Android-specific `Revoked` error now includes the revocation list entry that indicates a revocation or suspension.
-
 * **Fixes**
     * **Fix a glaring signature verification bug in the Supreme `AttestationVerifier`**
         * **Neither Makoto nor Roboto were ever affected by this**
@@ -34,7 +33,6 @@ If you have been using Makoto and/or Roboto directly, you were never at risk
     * Allow specifying CSR attributes and extensions for fully integrated flows     
     * Fix nonce validity duration calculation
     * Additional fixes to exception equality checks
-
 * **API-only Changes**
     * Remove all deprecations marked for removal with 1.0.0.
     * Switch order of `androidattestationConfigurationJ` and `iosAttestationconfigurationJ` for Java-oriented Makoto constructor
@@ -53,7 +51,7 @@ If you have been using Makoto and/or Roboto directly, you were never at risk
     * Make attestation functions suspending and provide blocking wrappers under old `@JvmName`s
     * Make Makoto configs publicly accessible
     * Deprecate misnamed and overly complex APIs. They will be removed with 1.1
-
+    * Fail early on invalid iOS team identifier length
 * **Dependency and Build updates**
     * Gradle 9 + Kotlin 2.3
         * Return value checker defaults to `check`